src/backend/requirements.txt
# Core dependencies
qrcode>=7.3.1
pillow>=9.0.0
click>=8.0.0

# Development dependencies
pytest>=7.0.0
